third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/document_getAll_set.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: Sample Test</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

// Feel free to use async here. I'm rusty on the syntax, so I went with what I
// know in the interest of getting you unblocked quickly.
promise_test(testCase => {
  return cookieStore.set('answer', '42').then(() => {
    return cookieStore.getAll('answer').then(cookies => {
      console.log(cookies);
      assert_equals(cookies.length, 1);
      assert_equals(cookies[0].name, 'answer');
      assert_equals(cookies[0].value, '42');
    });
  });
}, 'cookieStore.getAll returns the cookie written by cookieStore.set');

</script>
